bait/demo/submodules/cate_suggestion/requirements.txt
absl-py==1.1.0
alabaster==0.7.12
autopep8==1.6.0
Babel==2.10.3
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
dgl-cu113==0.8.2
-f https://data.dgl.ai/wheels/repo.html
dglgo==0.0.1
docker-pycreds==0.4.0
docutils==0.18.1
filelock==3.7.1
fonttools==4.33.3
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.8.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
huggingface-hub==0.26.5
idna==3.3
imagesize==1.4.1
importlib-metadata==4.11.4
isort==5.10.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
networkx==2.8.4
nltk==3.7
numpy==1.23.0
numpydoc==1.4.0
oauthlib==3.2.0
packaging==21.3
pandas==1.4.4
pathtools==0.1.2
Pillow==9.1.1
promise==2.3
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.8.1
sentry-sdk==1.6.0
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smart-open==6.0.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.20.3
toml==0.10.2
torch==1.11.0+cu113
-f https://download.pytorch.org/whl/cu113/torch_stable.html
tqdm==4.64.0
transformers==4.46.3
typer==0.4.1
typing-extensions==4.2.0
urllib3==1.26.9
wandb==0.12.20
Werkzeug==2.1.2
zipp==3.8.0
unidecode
openpyxl
black
